I fixed a bug that was happening in the remove function.
The bug was:
when deleting a node, that was the first node of a level:
* -> 12 -> 15 -> *
the out put would show that:
12 is betwee 12 and 15
because it would check if there is no node before 12, then printed the
head.
but what we want, is if there is no node before 12, and head is pointing
to 12, then 12 is between * and 15

I just added a check condition of << head[i] && heads[i]->data != item
>>.

I tested it. Bug is fixed.

Signed-off-by: Hassan Attar <info@hassanattar.com>