I fixed a bug happening in representation
when you add an element to the beginning of the SkipList, it was by mistake showing that the rewiring is not happening correctly.
it would show that -1 is inserted after 2, which was due to the fact that it wasn't taking into consideration, the case where there is only one element in a level.

the fix was just checking the head to see if it is pointing to something, and whether that item is less or more than the item being inserted.

since it was just a problem with representation, I don't think more explanation will be relevant and necessary.

Signed-off-by: Hassan Attar <hattar00@mylangara.ca>