I have seperated this repo from the lean implementation of SkipList class.
This repo contains SkipListVerbose which is a verbose version of the skiplist class.
It inherits from the skipList class, and adds some extra methods to print information to the console.

the SkipList is added as a submodule to this repository, and it will be updated as SkipList class becomes more efficient and robust.

Signed-off-by: Hassan Attar <hattar00@mylangara.ca>